[FW][IMP] l10n_ar: show payment terms
In the Argentinean invoices report view, we were removing the Payment
Terms and we were only showing the Payment Term name in the header.
Now we add it back following the Odoo approach but we also keep the name
of the terms on the header.
We made this modification in the invoice report with and without payment.

original commit: 6a0cf2e

closes odoo/odoo#69998

X-original-commit: d4b6bd1a06bc2a3228147241ae0229a5ce4fe94e
Signed-off-by: Josse Colpaert <jco@openerp.com>