[FIX] web: no bounce button form field/label click
Since commit [1], a delay was added to "debounce" the quick edit to
check if the user did a click or a select.
Because of this delay, the edit button now bounces when clicking on
a field or label.

In this commit, we check if the user clicks on a "quick editable"
field or a label to prevent bouncing.

[1] https://github.com/odoo/odoo/commit/d4d38a13d07696a1e404ce7df77d49a6f99dc211

closes odoo/odoo#68835

X-original-commit: 232aae1538621d1df618d5a530350c920e5b6e5d
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>