[IMP] website_event: always display 'filter by tags' in frontend
This commit removes the ability to enable/disable the "filter by tags"
on /event. Not need for it to be customizable, it should be displayed
by default. Indeed this is basics of event management and having options
for that adds noise to configuration.

Also, from now on, only published tag categories will be used to sort
events on the frontend, whatever the ACLs.

Task-2488019
PR odoo/odoo#68901