[FIX] stock: inventory adjustement for user
Fix the access for stock user to stock.quant.
Since the inventory adjustement feature has been moved on
the stock.quant. The user should be able to write on them to
make their count. However we kept that a stock user can't validate
an inventory nor modify quantities on product

closes odoo/odoo#69006

X-original-commit: 9be5d793245e6a48d33da6c76a29d3f43a968d05
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>