[IMP] sale_timesheet: extend project update
This commit adds information in project update based on sale_timesheet models.

In the Project Right Panel, a profitability section is added with data
from the profitability report.
In the project update description a profitability section is also added
with data from account analytic lines or profitability report.

PR : #68899

task-2393768