[FIX] web: centered event name while dragging in month mode
before this commit, the name of the event and hour is not
correctly centered while dragging the pill in month mode.

This commit fixes the issue by adding the CSS property.
after this commit, the name of the event and hour should be
correctly centered while dragging the pill in month mode.

TaskID-2506060

closes odoo/odoo#73553

X-original-commit: d17c8f251b46e80856a7e34953fa63d1b2fde514
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>