[FIX] website: fix top position of modal
Before this commit, the top position of a modal in website was always
the height of the main menu navbar. Even if the main menu navbar was
not displayed when the user is not logged.

That bug was introduced in f94d947580743fbd0f282e51b96a921dddc5d60b

After this commit, the top position of a modal is the height of the
main menu navbar but only if that navbar is displayed.

task-2507949

closes odoo/odoo#69391

X-original-commit: ffe578e180c324050b83b1f72e01dbd560b1e72d
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>